passport/server/requirements.txt
Django==1.10.5
django-cors-middleware==1.3.1
djangorestframework==3.5.4
django_phonenumber_field==1.1.0
